+++ Fiero/Views/Challenges/QRCode/ScanViewController.swift
@@ -178,24 +178,3 @@ extension ScanViewController: AVCaptureVideoDataOutputSampleBufferDelegate {
 }
 
 public typealias ProcessQRCode = (_ data: Data) -> ()
-
-// MARK: UIDeviceOrientation and AVCaptureVideoOrientation Equivalent
-extension UIDeviceOrientation {
-    var forVideoOrientation: AVCaptureVideoOrientation {
-        switch self {
-        case .landscapeLeft:
-            return .landscapeRight
-        case .landscapeRight:
-            return .landscapeLeft
-        case .faceUp, .portrait:
-            return .portrait
-        case .faceDown, .portraitUpsideDown:
-            return .portraitUpsideDown
-        case .unknown:
-            return .portrait
-        @unknown default:
-            return .portrait
-        }
-    }
-}
-
